+++ EssentialFeed/EssentialFeedTests/FeedCache/XCTestCase+FailableInsertFeedStoreSpecs.swift
@@ -1,0 +1,29 @@
+//
+//  XCTestCase+FailableInsertFeedStoreSpecs.swift
+//  EssentialFeedTests
+//
+//  Created by Radosław Serek on 17/11/2022.
+//
+
+import EssentialFeed
+import XCTest
+
+extension FailableInsertFeedStoreSpecs where Self: XCTestCase {
+    func assertThatInsertDeliversFailureOnInsertionError(on sut: FeedStore, file: StaticString = #filePath, line: UInt = #line) {
+        let feed = uniqueImageFeed().local
+        let timestamp = Date()
+        
+        let insertionError = insert(feed: feed, timestamp: timestamp, with: sut)
+        
+        XCTAssertNotNil(insertionError, "Expected cache insertion to fail with an error", file: file, line: line)
+    }
+
+    func assertThatInsertHasNoSideEffectsOnInsertionError(on sut: FeedStore, file: StaticString = #filePath, line: UInt = #line) {
+        let feed = uniqueImageFeed().local
+        let timestamp = Date()
+        
+        insert(feed: feed, timestamp: timestamp, with: sut)
+        
+        expect(sut, toRetrieve: .empty, file: file, line: line)
+    }
+}
